Python_GPT_RAG/requirements.txt
langchain==0.1.20
langchain-openai==0.1.8
langchain-community==0.0.38
langchain-huggingface==0.0.16
sentence-transformers==2.3.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
